# 003-APIRESTOpenAPI/MiClienteReservas/api/openapi.yaml
openapi: 3.0.4
info:
  title: DemoApi
  version: "1.0"
servers:
- url: /
security:
- Bearer: []
...
paths:
...
  /api/reservas:
    get:
      operationId: GetReservas
      parameters:
      - explode: true
        in: query
        name: fecha
        required: false
        schema:
          format: date
          type: string
        style: form
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaListApiResponse"
          description: OK
      summary: Obtiene todas las reservas o las filtra por fecha (yyyy-MM-dd).
      tags:
      - Reserva
    post:
      operationId: CreateReserva
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/CreateReservaRequestDto"
              description: Datos necesarios para registrar una reserva.
        required: true
      responses:
        "401":
          description: Unauthorized
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Conflict
      tags:
      - Reserva
  /api/reservas/{id}:
    delete:
      operationId: DeleteReserva
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      - explode: true
        in: query
        name: eliminar
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: OK
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Not Found
      summary: Cancela una reserva (estado Cancelada). Con ?eliminar=true la elimina
        definitivamente.
      tags:
      - Reserva
    get:
      operationId: GetReservaById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Not Found
      tags:
      - Reserva
    put:
      operationId: UpdateReserva
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/UpdateReservaRequestDto"
              description: Datos que pueden modificarse de una reserva existente.
        required: true
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaApiResponse"
          description: Conflict
      tags:
      - Reserva
components:
  schemas:
...
    CreateReservaRequestDto:
      additionalProperties: false
      description: Datos necesarios para registrar una reserva.
      example:
        nombreSala: nombreSala
        fechaReserva: 2000-01-23
        horaInicio: horaInicio
        horaFin: horaFin
        nombreResponsable: nombreResponsable
        cantidadAsistentes: 81
        motivo: motivo
      properties:
        nombreSala:
          description: Nombre de la sala a reservar.
          maxLength: 100
          minLength: 2
          type: string
        fechaReserva:
          description: Fecha de la reserva (yyyy-MM-dd).
          format: date
          type: string
        horaInicio:
          description: Hora de inicio (HH:mm:ss).
          format: time
          type: string
        horaFin:
          description: Hora de fin (HH:mm:ss). Debe ser posterior a la hora de inicio.
          format: time
          type: string
        nombreResponsable:
          description: Nombre del responsable de la reserva.
          maxLength: 150
          minLength: 2
          type: string
        cantidadAsistentes:
          description: Cantidad de asistentes. Debe ser mayor a cero.
          format: int32
          maximum: 1000
          minimum: 1
          type: integer
        motivo:
          description: Motivo de la reunión.
          maxLength: 300
          minLength: 3
          type: string
      required:
      - fechaReserva
      - horaFin
      - horaInicio
      - motivo
      - nombreResponsable
      - nombreSala
      type: object
    EstadoReserva:
      description: Estados posibles de una reserva.
      enum:
      - Pendiente
      - Confirmada
      - Cancelada
      type: string
...
    Reserva:
      additionalProperties: false
      description: Representa la reserva de una sala de reuniones.
      example:
        id: 0
        nombreSala: nombreSala
        fechaReserva: 2000-01-23
        horaInicio: horaInicio
        horaFin: horaFin
        nombreResponsable: nombreResponsable
        cantidadAsistentes: 6
        motivo: motivo
        estado: ""
      properties:
        id:
          description: Identificador único de la reserva.
          format: int32
          type: integer
        nombreSala:
          description: Nombre de la sala reservada.
          nullable: true
          type: string
        fechaReserva:
          description: Fecha de la reserva.
          format: date
          type: string
        horaInicio:
          description: Hora de inicio de la reunión.
          format: time
          type: string
        horaFin:
          description: Hora de fin de la reunión.
          format: time
          type: string
        nombreResponsable:
          description: Nombre de la persona responsable de la reserva.
          nullable: true
          type: string
        cantidadAsistentes:
          description: Cantidad de asistentes a la reunión.
          format: int32
          type: integer
        motivo:
          description: Motivo de la reunión.
          nullable: true
          type: string
        estado:
          allOf:
          - $ref: "#/components/schemas/EstadoReserva"
          description: Estado actual de la reserva.
      type: object
    ReservaApiResponse:
      additionalProperties: false
      example:
        isSuccess: true
        message: message
        data: ""
        errors:
        - errors
        - errors
      properties:
        isSuccess:
          type: boolean
        message:
          nullable: true
          type: string
        data:
          allOf:
          - $ref: "#/components/schemas/Reserva"
          description: Representa la reserva de una sala de reuniones.
          nullable: true
        errors:
          items:
            type: string
          nullable: true
          type: array
      type: object
    ReservaListApiResponse:
      additionalProperties: false
      example:
        isSuccess: true
        message: message
        data:
        - id: 0
          nombreSala: nombreSala
          fechaReserva: 2000-01-23
          horaInicio: horaInicio
          horaFin: horaFin
          nombreResponsable: nombreResponsable
          cantidadAsistentes: 6
          motivo: motivo
          estado: ""
        - id: 0
          nombreSala: nombreSala
          fechaReserva: 2000-01-23
          horaInicio: horaInicio
          horaFin: horaFin
          nombreResponsable: nombreResponsable
          cantidadAsistentes: 6
          motivo: motivo
          estado: ""
        errors:
        - errors
        - errors
      properties:
        isSuccess:
          type: boolean
        message:
          nullable: true
          type: string
        data:
          items:
            $ref: "#/components/schemas/Reserva"
          nullable: true
          type: array
        errors:
          items:
            type: string
          nullable: true
          type: array
      type: object
...
    UpdateReservaRequestDto:
      additionalProperties: false
      description: Datos que pueden modificarse de una reserva existente.
      example:
        nombreSala: nombreSala
        fechaReserva: 2000-01-23
        horaInicio: horaInicio
        horaFin: horaFin
        nombreResponsable: nombreResponsable
        cantidadAsistentes: 81
        motivo: motivo
        estado: ""
      properties:
        nombreSala:
          description: Nuevo nombre de la sala.
          maxLength: 100
          minLength: 2
          type: string
        fechaReserva:
          description: Nueva fecha de la reserva (yyyy-MM-dd).
          format: date
          type: string
        horaInicio:
          description: Nueva hora de inicio (HH:mm:ss).
          format: time
          type: string
        horaFin:
          description: Nueva hora de fin (HH:mm:ss). Debe ser posterior a la hora
            de inicio.
          format: time
          type: string
        nombreResponsable:
          description: Nuevo nombre del responsable.
          maxLength: 150
          minLength: 2
          type: string
        cantidadAsistentes:
          description: Nueva cantidad de asistentes. Debe ser mayor a cero.
          format: int32
          maximum: 1000
          minimum: 1
          type: integer
        motivo:
          description: Nuevo motivo de la reunión.
          maxLength: 300
          minLength: 3
          type: string
        estado:
          allOf:
          - $ref: "#/components/schemas/EstadoReserva"
          description: "Nuevo estado de la reserva (Pendiente, Confirmada o Cancelada)."
          format: Enumeration
      required:
      - estado
      - fechaReserva
      - horaFin
      - horaInicio
      - motivo
      - nombreResponsable
      - nombreSala
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      description: "Escribe: Bearer {token}"
      scheme: bearer
      type: http